Fix: CI Python Version (#67)
It turns out that Hatch does **not** default to the Python version of the current shell. This means that our CI was always defaulting to the version of Python that Hatch was installed with, which at the current time is always `3.12`. This was completely silent, until some weird caching things broke.

This PR sets the `HATCH_PYTHON` environment variable to the output of the `actions/setup-python` step, so that Hatch uses the Python version that we want.